Beautified file, fixed some warnings + Small changes to text
I've formatted this file using the Prettier extension in VS Code. I tried not changing that much the order of the scripts (in the head), as you said. I added some comments in the head so it's easier to see things. I've added supplementary text to tell the user on which page they are.